move auth to its own class
