+++ xmtp-inbox-ios/ContentView.swift
@@ -8,26 +8,6 @@
 import SwiftUI
 import XMTP
 
-class Auth: ObservableObject {
-
-    enum AuthStatus {
-        case loadingKeys, signedOut, tryingDemo, connecting, connected(Client)
-    }
-
-    @Published var status: AuthStatus = .loadingKeys
-
-    func signOut() {
-        do {
-            try Keystore.deleteKeys()
-            withAnimation {
-                self.status = .signedOut
-            }
-        } catch {
-            print("Error signing out: \(error.localizedDescription)")
-        }
-    }
-}
-
 struct ContentView: View {
 
     @StateObject private var auth = Auth()
